avoid crash if CDPATH is set (Closes: #848693)
If, like me, you had the clever idea of using CDPATH to save time, the
behaviour of the "cd" command changed, and it now outputs the
directory it CDs in to. So this line 302:

    DEBIRF_BUILDD=$(cd "$DEBIRF_BUILDD" && pwd)

unexpectly yields:

/home/anarcat/src/stressant
/home/anarcat/src/stressant/.fakeroot-state.debirf-stressant

in my tests here.

i would simply silence the output of cd using > /dev/null, but since
this pattern is used in a few places, it's probably more reliable to
just unset CDPATH.