Ship /bin/unxz in the outer initramfs
Commit 65a7bbe switched from gzip to xz for the nested initramfs. This
included a change in the embedded shell script to call "unxz" to unpack
rootfs.cxz. However, no unxz is to be found in $PATH, as an /bin/unxz
(busybox symlink) was not shipped, unlike /bin/gunzip that was already
there.

As far as I understand it, this means that all builds have been broken
since that commit. I'm unsure how this has escaped attention thus far.