Use crc32 for xz, unbreaking "stupid simple" mode
Commit 65a7bbe switched the rootfs compression method, in both nested
and "stupid simple" modes, from gzip to xz.

However, xz's integrity check type defaults to crc64, but Linux (as of
v6.1) is unable to unpack it, as it only supports none and crc32. This
does not matter in the default, nested mode where the inner rootfs is
unpacked by busybox, and the outer one (unpacked by Linux) is compressed
with gzip. It does matter for the non-default stupid simple mode though,
which simply does not work right now.

Pass --check=crc32 to xz when packing the initramfs. As this should have
a minimal impact and be inconsequential to the nested builds, do so
unconditionally.